Add (unimplemented) coalesce_available_sections
This method will allow you to combine free sections within the bin.

When working under a tight performance budget you have the ability
to distribute this coalescing over multiple calculations.

I don't need it quite yet, so I'll implement it later. But exposing
it now in case someone else needs it and wants to implement it.